pdf_conversion_agent.py
import fitz  # PyMuPDF
from pdfminer.high_level import extract_text
import openai
import argparse
import os

def main():
    parser = argparse.ArgumentParser(description="508 Compliant PDF Conversion Agent")
    parser.add_argument("input", help="Path to the input PDF file")
    parser.add_argument("output", help="Path to save the compliant PDF file")
    args = parser.parse_args()

    # TODO: Implement PDF conversion logic

if __name__ == "__main__":
    main()
