requirements.txt
PyMuPDF==1.24.9
pdfminer.six==20231228
openai==1.3.7
